Remove server-side Sentry, consolidate on PostHog
- Delete sentry.ts and remove @sentry/node dependency
- Replace all Sentry captureException/metrics calls with PostHog
- Add telemetry opt-out: DO_NOT_TRACK=1 env var and telemetry_enabled setting
- Gate telemetry on production environment
- Fix require() to use ESM import in posthog.ts
