+++ apps/server/package.json
@@ -27,7 +27,6 @@
     "@hono/node-server": "^2.0.4",
     "@hono/zod-validator": "^0.8.0",
     "@modelcontextprotocol/sdk": "^1.29.0",
-    "@sentry/node": "^10.53.1",
     "ai": "^6.0.191",
     "hono": "^4.12.22",
     "hono-rate-limiter": "^0.5.3",
