+++ apps/server/src/routes/transcribe.ts
@@ -1,12 +1,8 @@
 import { Hono } from "hono";
 import { getDb } from "../lib/db.js";
 import { postProcess } from "../lib/post-process.js";
-import {
-  capture,
-  captureException as posthogCaptureException,
-} from "../lib/posthog.js";
+import { capture, captureException } from "../lib/posthog.js";
 import { getDefaultModels } from "../lib/providers.js";
-import { captureException, metrics } from "../lib/sentry.js";
 import { getProvider } from "../lib/streaming/registry.js";
 import { getApiKeyForProvider } from "../lib/streaming-stt.js";
 import { resolveAsrVocabularyBias } from "../lib/vocabulary-bias.js";
@@ -103,13 +99,9 @@ const transcribeRoute = new Hono().post("/", async (c) => {
       );
     }
   } catch (err) {
-    captureException(err);
-    posthogCaptureException(err, {
+    captureException(err, {
       provider: defaults.voice.provider,
       model: defaults.voice.model_id,
-    });
-    metrics.count("transcription.error", 1, {
-      attributes: { provider: defaults.voice.provider },
     });
     capture("transcription failed", {
       provider: defaults.voice.provider,
@@ -126,22 +118,6 @@ const transcribeRoute = new Hono().post("/", async (c) => {
   }
 
   const durationMs = Date.now() - start;
-
-  const tags = {
-    provider: defaults.voice.provider,
-    model: defaults.voice.model_id,
-  };
-  metrics.count("transcription.count", 1, { attributes: tags });
-  metrics.distribution("transcription.latency", durationMs, {
-    unit: "millisecond",
-    attributes: tags,
-  });
-  if (audioDurationMs > 0) {
-    metrics.distribution("transcription.audio_duration", audioDurationMs, {
-      unit: "millisecond",
-      attributes: tags,
-    });
-  }
 
   if (!rawText.trim()) {
     return c.json({
